% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cell_weight.R
\name{cell_weight}
\alias{cell_weight}
\title{Perform cell weighting.}
\usage{
cell_weight(sample_df, attribute_col, population_cells)
}
\arguments{
\item{sample_df}{dataframe}

\item{attribute_col}{string}

\item{population_cells}{list of format list(joint_attribute1=x, joint_attribute2=y, ...)}
}
\value{
weighted_sample (dataframe)
}
\description{
Requires full joint distribution of covariates of interest.
}
