% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coefplot.r
\name{coefplot_mlm}
\alias{coefplot_mlm}
\title{Coefficient plot for mixed-effects model}
\usage{
coefplot_mlm(model, intercept = FALSE, sort = FALSE,
  title = "Predictors", subtitle = "",
  caption = "Bars represent 95 \% C.I.", labels = NULL,
  conf.level = 0.95, nudge_x = 0, nudge_y = 0.3)
}
\arguments{
\item{nudge_y}{}
}
\description{
Coefficient plot for mixed-effects model
}
