% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{combine_sample_synthetic}
\alias{combine_sample_synthetic}
\title{Combine sample and synthetic dataframes into one for modeling.}
\usage{
combine_sample_synthetic(sample_df, synthetic_population,
  factor_cols = NA, predictor_cols = NA)
}
\arguments{
\item{sample_df}{dataframe}

\item{synthetic_population}{dataframe}

\item{factor_cols}{vector}

\item{predictor_cols}{vector}
}
\value{
model_df dataframe
}
\description{
Combine sample and synthetic dataframes into one for modeling.
}
