% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/confidence_interval.r
\name{confidence_interval}
\alias{confidence_interval}
\title{Get confidence interval for mean, standard devation and number of
observations}
\usage{
confidence_interval(m, s, n, conf.int = 0.95)
}
\arguments{
\item{m}{mean.}

\item{s}{standard deviation of the mean.}

\item{n}{number of observations.}

\item{conf.int}{confidence interval (defaults to 0.95).}
}
\value{
a list with mean, sd, error, confidence interval, lower bound of
confidence interval (lower), and upper bound of confidence interval (upper).
}
\description{
Get confidence interval for mean, standard devation and number of
observations
}
