% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/corr_histogram.r
\name{corr_histogram}
\alias{corr_histogram}
\title{Create histogram of correlations}
\usage{
corr_histogram(df, use = "complete.obs")
}
\arguments{
\item{df}{data frame with numeric variables that should be correlated.}

\item{use}{(defalt is "complete.obs").}
}
\value{
a ggplot2 plot with a histogram with correlations.
}
\description{
Create histogram of correlations between variables in a data frame.
}
