% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{create_comparison_df}
\alias{create_comparison_df}
\title{Create wide dataframe to compare results of various methods.
Intended to be used to simplify tutorial notebook.}
\usage{
create_comparison_df(df_list, attribute)
}
\arguments{
\item{df_list}{named list of dataframes of style list(col_name=sample_distribution_df, ...)}

\item{attribute}{string}
}
\value{
comparison_df dataframe
}
\description{
Create wide dataframe to compare results of various methods.
Intended to be used to simplify tutorial notebook.
}
