% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crosstab.r
\name{crosstab}
\alias{crosstab}
\title{Cross tabulation  of two variables}
\usage{
crosstab(x, y, digits = 1, na.rm = FALSE, labels = TRUE)
}
\arguments{
\item{x}{vector.}

\item{y}{vector (grouping variable).}

\item{digits}{round percentage to this number of digits.}

\item{na.rm}{a logical value indicating whether NA values should be
stripped before the computation proceeds.}
}
\value{
a tibble is returned if the input (`x`) is a vector. If the input is
a data frame, then frequencies for each column is printed on the screen.
}
\description{
Prints or returns a frequency table from a vector or a data frame.
}
\examples{
\donotrun{
crosstab(iris$Petal.Width, iris$Species)
}
}
