% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/entropy_balance.R
\name{entropy_balance}
\alias{entropy_balance}
\title{Perform entropy balancing. TODO.}
\usage{
entropy_balance(sample_df)
}
\arguments{
\item{sample_df}{dataframe}
}
\value{
weighred_sample (dataframe)
}
\description{
Perform entropy balancing. TODO.
}
