% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_corr_code.r
\name{generate_corr_code}
\alias{generate_corr_code}
\title{Generate R code that creates a correlation matrix}
\usage{
generate_corr_code(x, digits = NULL, early_line_break = FALSE,
  method = "cor", target_variable = "df")
}
\arguments{
\item{x}{a data frame with numeric variables, or a matrix with correlations
or covariances.}

\item{digits}{number of decimal digits. Default (NULL) will show all
available digits as specified by R options.}

\item{early_line_break}{use early line break so that each value is shown
on a single line instead of several. Defaults to FALSE.}

\item{method}{method for creating the matrix, either "cor" for correlations
(default) or "cov" for covariances. This argument is only used when}

\item{target_variable}{target variable name for the code. Defaults to "df".
"x" is a data frame.}
}
\value{
R code that reproduce the correlation matrix of the data frame.
}
\description{
Generate R code from a data frame or matrix.
This is useful, for example, when you need to create a correlation
matrix that was used by a structural equation model for reproducable
research.
}
\details{
You supply a data frame with numeric variables (or a matrix with
correlation/covariance). The function then creates code that
can reproduce the data as a correlation matrix or a covariance matrix.
}
\examples{
\dontrun{
# Create random data.
set.seed(5543)
data <- data.frame(x = rnorm(100),
                   y = rnorm(100),
                   z = rnorm(100),
                   w = rnorm(100),
                   q = rnorm(100))

# Generate R code for the data.
cat(generate_corr_code(data))

# Generate R code for the data, with 3 decimal digits.
cat(generate_corr_code(data, digits=3))

# Generate the same code, but keep each value on one line.
cat(generate_corr_code(data, digits=3, early_line_break=TRUE))

# Save R code to a file.
code <- generate_corr_code(data)
fileConn <- file("create_data.r")
writeLines(code, fileConn)
close(fileConn)
}
}
