% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_corr_code.r
\name{generate_cov_code}
\alias{generate_cov_code}
\title{Code generator for reproducable covariance matrix}
\usage{
generate_cov_code(x, digits = NULL, early_line_break = FALSE,
  method = "cov", formula = "YOUR lavaan MODEL HERE",
  target_variable = "data")
}
\arguments{
\item{x}{a data frame with numeric variables, or a matrix with correlations
or covariances.}

\item{digits}{number of decimal digits. Default (NULL) will show all
available digits as specified by R options. The higher the number of
decimal digits, the accurate the reproducable model will be.}

\item{early_line_break}{use early line break so that each value is shown
on a single line instead of several. This is for cosmetic reasons
and is only of importance if you plan to print your code or include it in a
manuscript or article. Defaults to FALSE.}

\item{method}{method for creating the matrix, either "cor" for correlations
or "cov" for covariances (default). This argument is only used when argument
"x" is a data frame.}

\item{formula}{optional argument that specifies the lavaan formula that
should be included in the code.}

\item{target_variable}{target variable name for the code. Defaults to "data",
but can be anything you want.}
}
\value{
Character with the generated R code to reproduce the covariance
matrix and the necessary lavaan code to run it.
}
\description{
Generate R code from your data frame or an existing covariance matrix.
This is useful when you need to create a reproducable covariance matrix
that can be used by a structural equation model (SEM) in lavaan.
}
\details{
You supply a data frame with numeric variables (or a covariance matrix that
you have already prepared). The function then generates R code that
can reproduce the SEM model as a correlation matrix or a covariance matrix.
}
\examples{
\dontrun{
# Create random data.
set.seed(5543)
data <- data.frame(x = rnorm(100),
                   y = rnorm(100),
                   z = rnorm(100),
                   w = rnorm(100),
                   q = rnorm(100))

# Generate R code for the data.
cat(generate_cov_code(data))

# Generate R code for the data, with 5 decimal digits.
cat(generate_cov_code(data, digits=5))

# Generate the same code, but keep each value on one line for easier reading.
cat(generate_cov_code(data, digits=5, early_line_break=TRUE))

# Save R code to a file.
code <- generate_cov_code(data)
fileConn <- file("create_data.r")
writeLines(code, fileConn)
close(fileConn)
}
}
