% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/match.R
\name{get_most_similar}
\alias{get_most_similar}
\title{Return index of most similar column in distance matrix. 
Does not consider target column, breaks ties at random,
and only consider rows that have not been matched to yet.}
\usage{
get_most_similar(ix, mat, relevant_inds, used_inds)
}
\arguments{
\item{ix}{int}

\item{mat}{matrix}

\item{relevant_inds}{vector}

\item{used_inds}{vector}
}
\value{
original_col (int)
}
\description{
Return index of most similar column in distance matrix. 
Does not consider target column, breaks ties at random,
and only consider rows that have not been matched to yet.
}
