% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coefplot.r
\name{grouped_coefplot_mlm}
\alias{grouped_coefplot_mlm}
\title{Grouped coefficient plot for mixed-effects model}
\usage{
grouped_coefplot_mlm(model, intercept = FALSE, sort = FALSE,
  title = "Predictors", conf.level = 0.95,
  caption = "Bars represent 95 \% C.I.", labels = NULL,
  grouplabels = NULL, nudge_y = 0.3, nudge_x = 0)
}
\arguments{
\item{nudge_x}{}
}
\description{
Grouped coefficient plot for mixed-effects model
}
