% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/match.R
\name{match_sample}
\alias{match_sample}
\title{Perform matching.
Takes synthetic population dataset and matches 
sample cases to observations in that synthetic dataset.}
\usage{
match_sample(sample_df, synthetic_population, similarity_method = "rf",
  factor_cols = NA, predictor_cols = NA, quietly = FALSE)
}
\arguments{
\item{sample_df}{dataframe}

\item{synthetic_population}{dataframe}

\item{similarity_method}{string}

\item{factor_cols}{vector}

\item{predictor_cols}{vector}

\item{quietly}{boolean}
}
\value{
matched_sample (dataframe)'
}
\description{
Perform matching.
Takes synthetic population dataset and matches 
sample cases to observations in that synthetic dataset.
}
