% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalize.r
\name{normalize}
\alias{normalize}
\title{Normalize vector}
\usage{
normalize(x, na.rm = TRUE)
}
\arguments{
\item{x}{vector to normalize.}

\item{na.rm}{whether or not to remove NA values. Default is TRUE.}
}
\value{
a vector with the values normalized.
}
\description{
Normalize vector
}
\examples{
\donotrun{
x <- rnorm(50, 100)
normalize(x)
}
}
