% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot2_themes.r
\name{plot_theme_modifier_chart}
\alias{plot_theme_modifier_chart}
\title{Plot examples of all theme modifiers}
\usage{
plot_theme_modifier_chart()
}
\value{

}
\description{
Plot examples of all theme modifiers
}
