% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/propensity_weight.R
\name{propensity_weight}
\alias{propensity_weight}
\title{Perform propensity weighting.}
\usage{
propensity_weight(sample_df, synthetic_population, model = "logreg",
  factor_cols = NA, predictor_cols = NA)
}
\arguments{
\item{sample_df}{dataframe}

\item{synthetic_population}{dataframe}

\item{model}{string ('rf', 'logreg')}

\item{factor_cols}{vector}

\item{predictor_cols}{vector}
}
\value{
weighted_sample (dataframe)
}
\description{
Computes the inverse propensity that each sample observation
was included in the sample as opposed to being from the population.
}
