% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rake.R
\name{rake}
\alias{rake}
\title{Perform raking. If initialize_weights == FALSE, df must have a column named weight.}
\usage{
rake(sample_df, population_params, iterations = 1,
  initialize_weights = TRUE)
}
\arguments{
\item{sample_df}{dataframe}

\item{population_params}{list of lists of style list(attribute=list(val1=x, val2=y, ...), ...)}

\item{iterations}{int}

\item{initialize_weights}{boolean}
}
\value{
sample_df (dataframe)
}
\description{
Perform raking. If initialize_weights == FALSE, df must have a column named weight.
}
