% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/match.R
\name{random_forest_similarity}
\alias{random_forest_similarity}
\title{Compute similarity matrix from random forest.}
\usage{
random_forest_similarity(sample_df, synthetic_population, factor_cols,
  predictor_cols)
}
\arguments{
\item{synthetic_population}{dataframe}

\item{factor_cols}{vector}

\item{predictor_cols}{vector}

\item{:}{sample_df dataframe}
}
\value{
similarity_matrix (dataframe)
}
\description{
This method is described in this Pew article:
https://www.pewresearch.org/methods/2018/01/26/how-different-weighting-methods-work/.
}
