% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coefplot.r
\name{relabel}
\alias{relabel}
\title{Title}
\usage{
relabel(old_labels, new_labels)
}
\arguments{
\item{old_labels}{character vector with old labels.}

\item{new_labels}{character vector with new labels.}
}
\value{
character vector with the old labels replaced with the new labels.
}
\description{
Title
}
