% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reverse_code.r
\name{reverse_code}
\alias{reverse_code}
\title{Reverse code a numeric vector}
\usage{
reverse_code(x, minwarn = NULL, maxwarn = NULL)
}
\arguments{
\item{x}{a numeric vector.}

\item{minwarn}{a minimum threshold of vector where the function will break.}

\item{maxwarn}{a maximum threshold of vector where the function will break.}
}
\value{
a vector with values reversed.
}
\description{
Both minwarn and maxwarn are useful to constrain the variable to not
overflow its boundaries, and throw an error if it occurs.
}
