% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coefplot.r
\name{rotate_legend}
\alias{rotate_legend}
\title{Title}
\usage{
rotate_legend(gg)
}
\arguments{
\item{gg}{}
}
\description{
Solution by Sandy Muspratt:
https://stackoverflow.com/questions/40229263/how-to-rotate-legend-symbols-in-ggplot2?noredirect=1&lq=1
}
