% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot2_themes.r
\name{theme_boldandblue}
\alias{theme_boldandblue}
\title{Bold and blue ggplot2 theme}
\usage{
theme_boldandblue(font_size = 18, lines = TRUE, legend = TRUE, ...)
}
\arguments{
\item{...}{}
}
\value{

}
\description{
Bold and blue ggplot2 theme
}
