% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot2_themes.r
\name{theme_economist}
\alias{theme_economist}
\title{Economist style ggplot2 theme}
\usage{
theme_economist(font_size = 12, lines = TRUE, legend = TRUE, ...)
}
\arguments{
\item{...}{}
}
\value{

}
\description{
Economist style ggplot2 theme
}
