% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot2_themes.r
\name{theme_green}
\alias{theme_green}
\title{Greenish ggplot2 theme}
\usage{
theme_green(font_size = 12, lines = TRUE, legend = TRUE, ...)
}
\arguments{
\item{...}{}
}
\value{

}
\description{
Greenish ggplot2 theme
}
