% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot2_themes.r
\name{theme_pinkygray}
\alias{theme_pinkygray}
\title{Pinky gray ggplot2 theme}
\usage{
theme_pinkygray(font_size = 12, lines = TRUE, legend = TRUE, ...)
}
\arguments{
\item{...}{}
}
\value{

}
\description{
Pinky gray ggplot2 theme
}
