% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot2_themes.r
\name{theme_powerpoint}
\alias{theme_powerpoint}
\title{Larger ggplot2 theme for Powerpoint presentations}
\usage{
theme_powerpoint(font_size = 20, ...)
}
\arguments{
\item{font_size}{base font size.}

\item{...}{additional parameters passed to ggplot2 theme().}
}
\value{
ggplot2 theme.
}
\description{
Uses larger text size to make it easier to read.
}
\details{
Black text on white background.
}
