% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot2_themes.r
\name{theme_powerpoint_black}
\alias{theme_powerpoint_black}
\title{Black ggplot2 theme for Powerpoint presentations}
\usage{
theme_powerpoint_black(base_size = 12, base_family = NULL, ...)
}
\arguments{
\item{base_size}{base font size.}

\item{base_family}{base font family.}

\item{...}{additional parameters passed to ggplot2 theme().}
}
\value{
ggplot2 theme.
}
\description{
White text on black background.
}
\details{
Based on https://jonlefcheck.net/2013/03/11/black-theme-for-ggplot2-2/.
}
