% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot2_themes.r
\name{theme_publish}
\alias{theme_publish}
\title{ggplot2 theme for journal publications}
\usage{
theme_publish(base_family = NULL, ...)
}
\arguments{
\item{...}{additional parameters passed to ggplot2 theme().}
}
\value{
ggplot2 theme.
}
\description{
ggplot2 theme for journal publications
}
