% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot2_theme_modifiers.r
\name{theme_smaller}
\alias{theme_smaller}
\title{Modify existing theme by making the text size smaller}
\usage{
theme_smaller(font_size = 10, ...)
}
\arguments{
\item{font_size}{base font size of all text..}

\item{...}{additional parameters passed to ggplot2 theme.}
}
\value{
ggplot2 theme.
}
\description{
Modify existing theme by making the text size smaller
}
