% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot2_themes.r
\name{theme_stata}
\alias{theme_stata}
\title{Stata style ggplot2 theme}
\usage{
theme_stata(font_size = 12, lines = TRUE, legend = TRUE, ...)
}
\arguments{
\item{...}{}
}
\value{

}
\description{
Stata style ggplot2 theme
}
