% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot2_themes.r
\name{theme_template}
\alias{theme_template}
\title{Template for theme ggplot2 themes.}
\usage{
theme_template(
  font_size,
  font_family = "Arial",
  lines,
  legend,
  bg_color,
  panel_bg_color,
  title_color,
  subtitle_color,
  axis_color,
  secondary_text_color,
  ...
)
}
\arguments{
\item{font_size}{base font size.}

\item{lines}{whether to show lines or not.}

\item{legend}{whether to show legend or not.}

\item{theme}{}
}
\value{

}
\description{
Used internally. Not intended to be exported.
}
