% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coefplot.r
\name{tidy_model}
\alias{tidy_model}
\title{Tidy model and append labels, names, sort and remove intercept}
\usage{
tidy_model(
  model,
  name = NULL,
  intercept = FALSE,
  sort = FALSE,
  labels = NULL,
  conf.level = 0.95
)
}
\arguments{
\item{model}{a regression model}

\item{name}{name of the model.}

\item{intercept}{whether intercept should be included (default is FALSE).}

\item{sort}{whether estimates should be sorted in descending order.}

\item{labels}{vector of estimate label names. Must be the same length as
the number of estimates (including intercept).}
}
\value{

}
\description{
Tidy model and append labels, names, sort and remove intercept
}
